dt: Heavily rework as include file
From comments[1] on the LKML it seems that the devicetree
overlay/capemgr approach to loading cape configurations is put on hold
for a while. For now, I'll just use a .dtsi file. To build a
heater-bone-enabled devicetree, create a link to this file in the
arch/arm/boot/dts directory of the kernel tree and add the following to
arch/arm/boot/dts/am335x-boneblack.dts,

    #include <heater-bone.dtsi>

Also be sure to set `status = "disabled";` in the `hdmi` section.

[1] 1391608084-12665-1-git-send-email-ml@embed.me.uk